+++ pkg/scalers/redis_scaler_test.go
@@ -9,9 +9,13 @@ import (
 )
 
 var testRedisResolvedEnv = map[string]string{
-	"REDIS_HOST":     "none",
-	"REDIS_PORT":     "6379",
-	"REDIS_PASSWORD": "none",
+	"REDIS_HOST":              "none",
+	"REDIS_PORT":              "6379",
+	"REDIS_USERNAME":          "none",
+	"REDIS_PASSWORD":          "none",
+	"REDIS_SENTINEL_MASTER":   "none",
+	"REDIS_SENTINEL_USERNAME": "none",
+	"REDIS_SENTINEL_PASSWORD": "none",
 }
 
 type parseRedisMetadataTestData struct {
@@ -174,6 +178,115 @@ func TestParseRedisClusterMetadata(t *testing.T) {
 			},
 			wantErr: nil,
 		},
+		{
+			name: "username given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"username": "username",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "username",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "username given in metadata",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+				"username": "username",
+			},
+			authParams: map[string]string{},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "username",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "username given in metadata from env",
+			metadata: map[string]string{
+				"hosts":           "a, b, c",
+				"ports":           "1, 2, 3",
+				"listName":        "mylist",
+				"usernameFromEnv": "REDIS_USERNAME",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "none",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "password given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"password": "password",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					password:  "password",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "password given in metadata from env",
+			metadata: map[string]string{
+				"hosts":           "a, b, c",
+				"ports":           "1, 2, 3",
+				"listName":        "mylist",
+				"passwordFromEnv": "REDIS_PASSWORD",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					password:  "none",
+				},
+			},
+			wantErr: nil,
+		},
 	}
 
 	for _, testCase := range cases {
@@ -185,6 +298,411 @@ func TestParseRedisClusterMetadata(t *testing.T) {
 				AuthParams:      c.authParams,
 			}
 			meta, err := parseRedisMetadata(config, parseRedisClusterAddress)
+			if c.wantErr != nil {
+				assert.Contains(t, err.Error(), c.wantErr.Error())
+			} else {
+				assert.NoError(t, err)
+			}
+			assert.Equal(t, c.wantMeta, meta)
+		})
+	}
+}
+
+func TestParseRedisSentinelMetadata(t *testing.T) {
+	cases := []struct {
+		name        string
+		metadata    map[string]string
+		resolvedEnv map[string]string
+		authParams  map[string]string
+		wantMeta    *redisMetadata
+		wantErr     error
+	}{
+		{
+			name:     "empty metadata",
+			wantMeta: nil,
+			wantErr:  errors.New("no addresses or hosts given. address should be a comma separated list of host:port or set the host/port values"),
+		},
+		{
+			name: "unequal number of hosts/ports",
+			metadata: map[string]string{
+				"hosts": "a, b, c",
+				"ports": "1, 2",
+			},
+			wantMeta: nil,
+			wantErr:  errors.New("not enough hosts or ports given. number of hosts should be equal to the number of ports"),
+		},
+		{
+			name: "no list name",
+			metadata: map[string]string{
+				"hosts":      "a, b, c",
+				"ports":      "1, 2, 3",
+				"listLength": "5",
+			},
+			wantMeta: nil,
+			wantErr:  errors.New("no list name given"),
+		},
+		{
+			name: "invalid list length",
+			metadata: map[string]string{
+				"hosts":      "a, b, c",
+				"ports":      "1, 2, 3",
+				"listName":   "mylist",
+				"listLength": "invalid",
+			},
+			wantMeta: nil,
+			wantErr:  errors.New("list length parsing error"),
+		},
+		{
+			name: "address is defined in auth params",
+			metadata: map[string]string{
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"addresses": ":7001, :7002",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{":7001", ":7002"},
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "hosts and ports given in auth params",
+			metadata: map[string]string{
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"hosts": "   a, b,    c ",
+				"ports": "1, 2, 3",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "hosts and ports given in auth params",
+			metadata: map[string]string{
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"hosts": "   a, b,    c ",
+				"ports": "1, 2, 3",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "username given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"username": "username",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "username",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "username given in metadata",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+				"username": "username",
+			},
+			authParams: map[string]string{},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "username",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "username given in metadata from env",
+			metadata: map[string]string{
+				"hosts":           "a, b, c",
+				"ports":           "1, 2, 3",
+				"listName":        "mylist",
+				"usernameFromEnv": "REDIS_USERNAME",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					username:  "none",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "password given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"password": "password",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					password:  "password",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "password given in metadata from env",
+			metadata: map[string]string{
+				"hosts":           "a, b, c",
+				"ports":           "1, 2, 3",
+				"listName":        "mylist",
+				"passwordFromEnv": "REDIS_PASSWORD",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses: []string{"a:1", "b:2", "c:3"},
+					hosts:     []string{"a", "b", "c"},
+					ports:     []string{"1", "2", "3"},
+					password:  "none",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelUsername given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"sentinelUsername": "sentinelUsername",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:        []string{"a:1", "b:2", "c:3"},
+					hosts:            []string{"a", "b", "c"},
+					ports:            []string{"1", "2", "3"},
+					sentinelUsername: "sentinelUsername",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelUsername given in metadata",
+			metadata: map[string]string{
+				"hosts":            "a, b, c",
+				"ports":            "1, 2, 3",
+				"listName":         "mylist",
+				"sentinelUsername": "sentinelUsername",
+			},
+			authParams: map[string]string{},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:        []string{"a:1", "b:2", "c:3"},
+					hosts:            []string{"a", "b", "c"},
+					ports:            []string{"1", "2", "3"},
+					sentinelUsername: "sentinelUsername",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelUsername given in metadata from env",
+			metadata: map[string]string{
+				"hosts":                   "a, b, c",
+				"ports":                   "1, 2, 3",
+				"listName":                "mylist",
+				"sentinelUsernameFromEnv": "REDIS_SENTINEL_USERNAME",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:        []string{"a:1", "b:2", "c:3"},
+					hosts:            []string{"a", "b", "c"},
+					ports:            []string{"1", "2", "3"},
+					sentinelUsername: "none",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelPassword given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"sentinelPassword": "sentinelPassword",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:        []string{"a:1", "b:2", "c:3"},
+					hosts:            []string{"a", "b", "c"},
+					ports:            []string{"1", "2", "3"},
+					sentinelPassword: "sentinelPassword",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelPassword given in metadata from env",
+			metadata: map[string]string{
+				"hosts":                   "a, b, c",
+				"ports":                   "1, 2, 3",
+				"listName":                "mylist",
+				"sentinelPasswordFromEnv": "REDIS_SENTINEL_PASSWORD",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:        []string{"a:1", "b:2", "c:3"},
+					hosts:            []string{"a", "b", "c"},
+					ports:            []string{"1", "2", "3"},
+					sentinelPassword: "none",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelMaster given in authParams",
+			metadata: map[string]string{
+				"hosts":    "a, b, c",
+				"ports":    "1, 2, 3",
+				"listName": "mylist",
+			},
+			authParams: map[string]string{
+				"sentinelMaster": "sentinelMaster",
+			},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:      []string{"a:1", "b:2", "c:3"},
+					hosts:          []string{"a", "b", "c"},
+					ports:          []string{"1", "2", "3"},
+					sentinelMaster: "sentinelMaster",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelMaster given in metadata",
+			metadata: map[string]string{
+				"hosts":          "a, b, c",
+				"ports":          "1, 2, 3",
+				"listName":       "mylist",
+				"sentinelMaster": "sentinelMaster",
+			},
+			authParams: map[string]string{},
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:      []string{"a:1", "b:2", "c:3"},
+					hosts:          []string{"a", "b", "c"},
+					ports:          []string{"1", "2", "3"},
+					sentinelMaster: "sentinelMaster",
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "sentinelMaster given in metadata from env",
+			metadata: map[string]string{
+				"hosts":                 "a, b, c",
+				"ports":                 "1, 2, 3",
+				"listName":              "mylist",
+				"sentinelMasterFromEnv": "REDIS_SENTINEL_MASTER",
+			},
+			authParams:  map[string]string{},
+			resolvedEnv: testRedisResolvedEnv,
+			wantMeta: &redisMetadata{
+				targetListLength: 5,
+				listName:         "mylist",
+				connectionInfo: redisConnectionInfo{
+					addresses:      []string{"a:1", "b:2", "c:3"},
+					hosts:          []string{"a", "b", "c"},
+					ports:          []string{"1", "2", "3"},
+					sentinelMaster: "none",
+				},
+			},
+			wantErr: nil,
+		},
+	}
+
+	for _, testCase := range cases {
+		c := testCase
+		t.Run(c.name, func(t *testing.T) {
+			config := &ScalerConfig{
+				TriggerMetadata: c.metadata,
+				ResolvedEnv:     c.resolvedEnv,
+				AuthParams:      c.authParams,
+			}
+			meta, err := parseRedisMetadata(config, parseRedisSentinelAddress)
 			if c.wantErr != nil {
 				assert.Contains(t, err.Error(), c.wantErr.Error())
 			} else {